requirements.txt
appnope==0.1.0
decorator==4.0.9
dj-database-url==0.4.1
Django==1.9.5
django-bootstrap3==7.0.1
django-debug-toolbar==1.4
django-extensions==1.6.1
djangorestframework==3.3.3
fake-factory==0.5.7
gnureadline==6.3.3
gunicorn==19.4.5
ipykernel==4.3.1
ipython==4.1.2
ipython-genutils==0.1.0
ipywidgets==4.1.1
Jinja2==2.8
jsonschema==2.5.1
jupyter==1.0.0
jupyter-client==4.2.2
jupyter-console==4.1.1
jupyter-core==4.1.0
MarkupSafe==0.23
mistune==0.7.2
nbconvert==4.1.0
nbformat==4.0.1
notebook==4.1.0
path.py==8.1.2
pexpect==4.0.1
pickleshare==0.6
Pillow==3.2.0
psycopg2==2.6.1
ptyprocess==0.5.1
Pygments==2.1.3
python-dateutil==2.5.2
pyzmq==15.2.0
qtconsole==4.2.0
simplegeneric==0.8.1
six==1.10.0
sqlparse==0.1.19
terminado==0.6
tornado==4.3
traitlets==4.2.1
Werkzeug==0.11.5
whitenoise==3.0
